requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.1
matplotlib==3.10.6
numba==0.62.1
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pynndescent==0.5.13
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
umap-learn==0.5.9.post2
